requirements.txt
# local dev + testing
pandas>=2.0
pyarrow>=14.0
boto3>=1.34
pytest>=8.0
pytest-cov>=5.0
aws-cdk-lib>=2.130.0
constructs>=10.3.0